history context and transition fixes
History was always the global history, should probably merge Location and HistoryContext into the same thing (location is the state, history the way to change it).

Transitions were only working before because the updates were deferred for react@alpha D: The promise body got called after a full UI update, so we never got out of `history.transitioning` in the UI.

Now it’ll work for sync or async.