Add Link APIs for better suspense control
Compatibility with current API is preserved.

This is based on the idea of curi router [1] that allows also passing
function as children in the Link component that gets the navigating
state as argument.

I've also passed navigating in getProps and if you're interested in this
approach I think the arguments of getProps and children function could
be the same.

This new functionality enables usages like the one below:
```
<Link to="">
  {navigating => <div>Home {navigating && <Spinner />}</div>}
</Link>
```

[1] https://github.com/pshrmn/curi/blob/master/packages/react-dom/src/Link.tsx

NOTE: async/await demands further rollup/babel config so promises were
preferred here.