Support inline routes with 'stop' prop
The biggest complaint I get when recommending to friends is the requirement to wrap routes in components. This is just a quick idea I had to support inline routes like so:
```javascript
<Router>
  <div path="/" stop>
    <h2>Welcome</h2>
  </div>
  <div path="dashboard" stop>
    <h2>Dashboard</h2>
  </div>
</Router>
```

Personally I don't like the `stop` prop name and I'm not even sold on the implementation. It's an ugly prop, but it temporarily gets the point across. This would be awesome to support. Thoughts?