feat: add an as prop to render custom element
Note: I don't necessarily know how to document this, but happy to make
that change too.

This is in regards to gatsbyjs/gatsby#7546, where a custom element needs
to be passed (e.g. a bootstrap link) and this provides that
functionality.

The API is similar to something like styled-components, and this change
is a non-breaking, backwards compatible API.

Let me know if you have any questions!